Finished Level 1, renamed bot to Frenchie
