--- src/main/java/Duke.java
@@ -1,17 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-        /*String logo = " ____        _        \n"
-                + "|  _ \\ _   _| | _____ \n"
-                + "| | | | | | | |/ / _ \\\n"
-                + "| |_| | |_| |   <  __/\n"
-                + "|____/ \\__,_|_|\\_\\___|\n";
-        System.out.println("Hello from\n" + logo); */
-        String skeleton = "____________________________________________________________\n" +
-                " Hello! I'm Frenchie\n" +
-                " What can I do for you?\n" +
-                "____________________________________________________________\n" +
-                " Bye. Hope to see you again soon!\n" +
-                "____________________________________________________________\n";
-        System.out.println(skeleton);
-    }
-}
